rpi/requirements.txt
numpy==1.26.3
RPi.GPIO==0.7.1
scipy==1.11.4
spidev==3.6
